[mem_cache][9/N] refactor: move DSAIndexerPoolHost to pool_host.dsa (#35306)
